+++ pkg/provider/azure_wrap.go
@@ -21,21 +21,9 @@ import (
 	"net/http"
 	"regexp"
 	"strings"
-	"sync"
-	"time"
 
-	"github.com/Azure/azure-sdk-for-go/services/compute/mgmt/2022-08-01/compute"
-	"github.com/Azure/azure-sdk-for-go/services/network/mgmt/2022-07-01/network"
-
-	"k8s.io/apimachinery/pkg/types"
-	cloudprovider "k8s.io/cloud-provider"
-	"k8s.io/klog/v2"
-	"k8s.io/utils/pointer"
-
-	azcache "sigs.k8s.io/cloud-provider-azure/pkg/cache"
 	"sigs.k8s.io/cloud-provider-azure/pkg/consts"
 	"sigs.k8s.io/cloud-provider-azure/pkg/retry"
-	"sigs.k8s.io/cloud-provider-azure/pkg/util/deepcopy"
 )
 
 var (
@@ -63,323 +51,6 @@ func checkResourceExistsFromError(err *retry.Error) (bool, *retry.Error) {
 	}
 
 	return false, err
-}
-
-// getVirtualMachine calls 'VirtualMachinesClient.Get' with a timed cache
-// The service side has throttling control that delays responses if there are multiple requests onto certain vm
-// resource request in short period.
-func (az *Cloud) getVirtualMachine(nodeName types.NodeName, crt azcache.AzureCacheReadType) (vm compute.VirtualMachine, err error) {
-	vmName := string(nodeName)
-	cachedVM, err := az.vmCache.Get(vmName, crt)
-	if err != nil {
-		return vm, err
-	}
-
-	if cachedVM == nil {
-		klog.Warningf("Unable to find node %s: %v", nodeName, cloudprovider.InstanceNotFound)
-		return vm, cloudprovider.InstanceNotFound
-	}
-
-	return *(cachedVM.(*compute.VirtualMachine)), nil
-}
-
-func (az *Cloud) getRouteTable(crt azcache.AzureCacheReadType) (routeTable network.RouteTable, exists bool, err error) {
-	if len(az.RouteTableName) == 0 {
-		return routeTable, false, fmt.Errorf("route table name is not configured")
-	}
-
-	cachedRt, err := az.rtCache.GetWithDeepCopy(az.RouteTableName, crt)
-	if err != nil {
-		return routeTable, false, err
-	}
-
-	if cachedRt == nil {
-		return routeTable, false, nil
-	}
-
-	return *(cachedRt.(*network.RouteTable)), true, nil
-}
-
-func (az *Cloud) getPublicIPAddress(pipResourceGroup string, pipName string, crt azcache.AzureCacheReadType) (network.PublicIPAddress, bool, error) {
-	cached, err := az.pipCache.Get(pipResourceGroup, crt)
-	if err != nil {
-		return network.PublicIPAddress{}, false, err
-	}
-
-	pips := cached.(*sync.Map)
-	pip, ok := pips.Load(pipName)
-	if !ok {
-		// pip not found, refresh cache and retry
-		cached, err = az.pipCache.Get(pipResourceGroup, azcache.CacheReadTypeForceRefresh)
-		if err != nil {
-			return network.PublicIPAddress{}, false, err
-		}
-		pips = cached.(*sync.Map)
-		pip, ok = pips.Load(pipName)
-		if !ok {
-			return network.PublicIPAddress{}, false, nil
-		}
-	}
-
-	pip = pip.(*network.PublicIPAddress)
-	return *(deepcopy.Copy(pip).(*network.PublicIPAddress)), true, nil
-}
-
-func (az *Cloud) listPIP(pipResourceGroup string, crt azcache.AzureCacheReadType) ([]network.PublicIPAddress, error) {
-	cached, err := az.pipCache.Get(pipResourceGroup, crt)
-	if err != nil {
-		return nil, err
-	}
-	pips := cached.(*sync.Map)
-	var ret []network.PublicIPAddress
-	pips.Range(func(key, value interface{}) bool {
-		pip := value.(*network.PublicIPAddress)
-		ret = append(ret, *pip)
-		return true
-	})
-	return ret, nil
-}
-
-func (az *Cloud) getSubnet(virtualNetworkName string, subnetName string) (network.Subnet, bool, error) {
-	var rg string
-	if len(az.VnetResourceGroup) > 0 {
-		rg = az.VnetResourceGroup
-	} else {
-		rg = az.ResourceGroup
-	}
-
-	ctx, cancel := getContextWithCancel()
-	defer cancel()
-	subnet, err := az.SubnetsClient.Get(ctx, rg, virtualNetworkName, subnetName, "")
-	exists, rerr := checkResourceExistsFromError(err)
-	if rerr != nil {
-		return subnet, false, rerr.Error()
-	}
-
-	if !exists {
-		klog.V(2).Infof("Subnet %q not found", subnetName)
-	}
-	return subnet, exists, nil
-}
-
-func (az *Cloud) getAzureLoadBalancer(name string, crt azcache.AzureCacheReadType) (lb *network.LoadBalancer, exists bool, err error) {
-	cachedLB, err := az.lbCache.GetWithDeepCopy(name, crt)
-	if err != nil {
-		return lb, false, err
-	}
-
-	if cachedLB == nil {
-		return lb, false, nil
-	}
-
-	return cachedLB.(*network.LoadBalancer), true, nil
-}
-
-func (az *Cloud) getSecurityGroup(crt azcache.AzureCacheReadType) (network.SecurityGroup, error) {
-	nsg := network.SecurityGroup{}
-	if az.SecurityGroupName == "" {
-		return nsg, fmt.Errorf("securityGroupName is not configured")
-	}
-
-	securityGroup, err := az.nsgCache.GetWithDeepCopy(az.SecurityGroupName, crt)
-	if err != nil {
-		return nsg, err
-	}
-
-	if securityGroup == nil {
-		return nsg, fmt.Errorf("nsg %q not found", az.SecurityGroupName)
-	}
-
-	return *(securityGroup.(*network.SecurityGroup)), nil
-}
-
-func (az *Cloud) getPrivateLinkService(frontendIPConfigID *string, crt azcache.AzureCacheReadType) (pls network.PrivateLinkService, err error) {
-	cachedPLS, err := az.plsCache.GetWithDeepCopy(*frontendIPConfigID, crt)
-	if err != nil {
-		return pls, err
-	}
-	return *(cachedPLS.(*network.PrivateLinkService)), nil
-}
-
-func (az *Cloud) newVMCache() (azcache.Resource, error) {
-	getter := func(key string) (interface{}, error) {
-		// Currently InstanceView request are used by azure_zones, while the calls come after non-InstanceView
-		// request. If we first send an InstanceView request and then a non InstanceView request, the second
-		// request will still hit throttling. This is what happens now for cloud controller manager: In this
-		// case we do get instance view every time to fulfill the azure_zones requirement without hitting
-		// throttling.
-		// Consider adding separate parameter for controlling 'InstanceView' once node update issue #56276 is fixed
-		ctx, cancel := getContextWithCancel()
-		defer cancel()
-
-		resourceGroup, err := az.GetNodeResourceGroup(key)
-		if err != nil {
-			return nil, err
-		}
-
-		vm, verr := az.VirtualMachinesClient.Get(ctx, resourceGroup, key, compute.InstanceViewTypesInstanceView)
-		exists, rerr := checkResourceExistsFromError(verr)
-		if rerr != nil {
-			return nil, rerr.Error()
-		}
-
-		if !exists {
-			klog.V(2).Infof("Virtual machine %q not found", key)
-			return nil, nil
-		}
-
-		if vm.VirtualMachineProperties != nil &&
-			strings.EqualFold(pointer.StringDeref(vm.VirtualMachineProperties.ProvisioningState, ""), string(consts.ProvisioningStateDeleting)) {
-			klog.V(2).Infof("Virtual machine %q is under deleting", key)
-			return nil, nil
-		}
-
-		return &vm, nil
-	}
-
-	if az.VMCacheTTLInSeconds == 0 {
-		az.VMCacheTTLInSeconds = vmCacheTTLDefaultInSeconds
-	}
-	return azcache.NewTimedCache(time.Duration(az.VMCacheTTLInSeconds)*time.Second, getter, az.Config.DisableAPICallCache)
-}
-
-func (az *Cloud) newLBCache() (azcache.Resource, error) {
-	getter := func(key string) (interface{}, error) {
-		ctx, cancel := getContextWithCancel()
-		defer cancel()
-
-		lb, err := az.LoadBalancerClient.Get(ctx, az.getLoadBalancerResourceGroup(), key, "")
-		exists, rerr := checkResourceExistsFromError(err)
-		if rerr != nil {
-			return nil, rerr.Error()
-		}
-
-		if !exists {
-			klog.V(2).Infof("Load balancer %q not found", key)
-			return nil, nil
-		}
-
-		return &lb, nil
-	}
-
-	if az.LoadBalancerCacheTTLInSeconds == 0 {
-		az.LoadBalancerCacheTTLInSeconds = loadBalancerCacheTTLDefaultInSeconds
-	}
-	return azcache.NewTimedCache(time.Duration(az.LoadBalancerCacheTTLInSeconds)*time.Second, getter, az.Config.DisableAPICallCache)
-}
-
-func (az *Cloud) newNSGCache() (azcache.Resource, error) {
-	getter := func(key string) (interface{}, error) {
-		ctx, cancel := getContextWithCancel()
-		defer cancel()
-		nsg, err := az.SecurityGroupsClient.Get(ctx, az.SecurityGroupResourceGroup, key, "")
-		exists, rerr := checkResourceExistsFromError(err)
-		if rerr != nil {
-			return nil, rerr.Error()
-		}
-
-		if !exists {
-			klog.V(2).Infof("Security group %q not found", key)
-			return nil, nil
-		}
-
-		return &nsg, nil
-	}
-
-	if az.NsgCacheTTLInSeconds == 0 {
-		az.NsgCacheTTLInSeconds = nsgCacheTTLDefaultInSeconds
-	}
-	return azcache.NewTimedCache(time.Duration(az.NsgCacheTTLInSeconds)*time.Second, getter, az.Config.DisableAPICallCache)
-}
-
-func (az *Cloud) newRouteTableCache() (azcache.Resource, error) {
-	getter := func(key string) (interface{}, error) {
-		ctx, cancel := getContextWithCancel()
-		defer cancel()
-		rt, err := az.RouteTablesClient.Get(ctx, az.RouteTableResourceGroup, key, "")
-		exists, rerr := checkResourceExistsFromError(err)
-		if rerr != nil {
-			return nil, rerr.Error()
-		}
-
-		if !exists {
-			klog.V(2).Infof("Route table %q not found", key)
-			return nil, nil
-		}
-
-		return &rt, nil
-	}
-
-	if az.RouteTableCacheTTLInSeconds == 0 {
-		az.RouteTableCacheTTLInSeconds = routeTableCacheTTLDefaultInSeconds
-	}
-	return azcache.NewTimedCache(time.Duration(az.RouteTableCacheTTLInSeconds)*time.Second, getter, az.Config.DisableAPICallCache)
-}
-
-func (az *Cloud) newPIPCache() (azcache.Resource, error) {
-	getter := func(key string) (interface{}, error) {
-		ctx, cancel := getContextWithCancel()
-		defer cancel()
-
-		pipResourceGroup := key
-		pipList, rerr := az.PublicIPAddressesClient.List(ctx, pipResourceGroup)
-		if rerr != nil {
-			return nil, rerr.Error()
-		}
-
-		pipMap := &sync.Map{}
-		for _, pip := range pipList {
-			pip := pip
-			pipMap.Store(pointer.StringDeref(pip.Name, ""), &pip)
-		}
-		return pipMap, nil
-	}
-
-	if az.PublicIPCacheTTLInSeconds == 0 {
-		az.PublicIPCacheTTLInSeconds = publicIPCacheTTLDefaultInSeconds
-	}
-	return azcache.NewTimedCache(time.Duration(az.PublicIPCacheTTLInSeconds)*time.Second, getter, az.Config.DisableAPICallCache)
-}
-
-func (az *Cloud) newPLSCache() (azcache.Resource, error) {
-	// for PLS cache, key is LBFrontendIPConfiguration ID
-	getter := func(key string) (interface{}, error) {
-		ctx, cancel := getContextWithCancel()
-		defer cancel()
-		plsList, err := az.PrivateLinkServiceClient.List(ctx, az.PrivateLinkServiceResourceGroup)
-		exists, rerr := checkResourceExistsFromError(err)
-		if rerr != nil {
-			return nil, rerr.Error()
-		}
-
-		if exists {
-			for i := range plsList {
-				pls := plsList[i]
-				if pls.PrivateLinkServiceProperties == nil {
-					continue
-				}
-				fipConfigs := pls.PrivateLinkServiceProperties.LoadBalancerFrontendIPConfigurations
-				if fipConfigs == nil {
-					continue
-				}
-				for _, fipConfig := range *fipConfigs {
-					if strings.EqualFold(*fipConfig.ID, key) {
-						return &pls, nil
-					}
-				}
-
-			}
-		}
-
-		klog.V(2).Infof("No privateLinkService found for frontendIPConfig %q", key)
-		plsNotExistID := consts.PrivateLinkServiceNotExistID
-		return &network.PrivateLinkService{ID: &plsNotExistID}, nil
-	}
-
-	if az.PlsCacheTTLInSeconds == 0 {
-		az.PlsCacheTTLInSeconds = plsCacheTTLDefaultInSeconds
-	}
-	return azcache.NewTimedCache(time.Duration(az.PlsCacheTTLInSeconds)*time.Second, getter, az.Config.DisableAPICallCache)
 }
 
 func (az *Cloud) useStandardLoadBalancer() bool {
@@ -425,30 +96,4 @@ func ConvertResourceGroupNameToLower(resourceID string) (string, error) {
 
 	resourceGroup := matches[1]
 	return strings.Replace(resourceID, resourceGroup, strings.ToLower(resourceGroup), 1), nil
-}
-
-// isBackendPoolOnSameLB checks whether newBackendPoolID is on the same load balancer as existingBackendPools.
-// Since both public and internal LBs are supported, lbName and lbName-internal are treated as same.
-// If not same, the lbName for existingBackendPools would also be returned.
-func isBackendPoolOnSameLB(newBackendPoolID string, existingBackendPools []string) (bool, string, error) {
-	matches := backendPoolIDRE.FindStringSubmatch(newBackendPoolID)
-	if len(matches) != 2 {
-		return false, "", fmt.Errorf("new backendPoolID %q is in wrong format", newBackendPoolID)
-	}
-
-	newLBName := matches[1]
-	newLBNameTrimmed := strings.TrimSuffix(newLBName, consts.InternalLoadBalancerNameSuffix)
-	for _, backendPool := range existingBackendPools {
-		matches := backendPoolIDRE.FindStringSubmatch(backendPool)
-		if len(matches) != 2 {
-			return false, "", fmt.Errorf("existing backendPoolID %q is in wrong format", backendPool)
-		}
-
-		lbName := matches[1]
-		if !strings.EqualFold(strings.TrimSuffix(lbName, consts.InternalLoadBalancerNameSuffix), newLBNameTrimmed) {
-			return false, lbName, nil
-		}
-	}
-
-	return true, "", nil
 }
